+++ version/changelog.md
@@ -7,6 +7,7 @@
 - navigate to parent folder automatically if folder does not exist (any more)
 - execute a command in current directory
 - copy path to clipboard
+- add feature for favorites in drive selection overview
 
 ### v1.0.0 (25.01.2026)